Tell Rubocop to use Ruby 2.3
For now, just to get Rubocop working again, I'm hard-coding `2.3` in here as it looks like the API for this method in Rubocop has changed to take a second argument that the plugin isn't supplying.

This should ultimately be replaced by something smarter that actually inspects `.rubocop.yml` to figure out the target Ruby version, though I think we'll still need some way to set a default if there isn't a value set in that file.